Transferred data from App.js to PostPage.js and imported PostPage to App
